Renamed the Project, Folders and namespaces
