+++ VRMotionCompensation.sln
@@ -1,0 +1,56 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29102.190
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "client_overlay", "client_overlay\client_overlay.vcxproj", "{33E075DB-922D-3252-976E-46B5721DC3DE}"
+	ProjectSection(ProjectDependencies) = postProject
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64} = {05AC9994-2B63-4DE5-ABF3-95CE346F3A64}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "driver_vrmotioncompensation", "driver_vrmotioncompensation\driver_vrmotioncompensation.vcxproj", "{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64} = {05AC9994-2B63-4DE5-ABF3-95CE346F3A64}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "lib_vrmotioncompensation", "lib_vrmotioncompensation\lib_vrmotioncompensation.vcxproj", "{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Debug|x64.ActiveCfg = Debug|x64
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Debug|x64.Build.0 = Debug|x64
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Debug|x86.ActiveCfg = Debug|x64
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Release|x64.ActiveCfg = Release|x64
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Release|x64.Build.0 = Release|x64
+		{33E075DB-922D-3252-976E-46B5721DC3DE}.Release|x86.ActiveCfg = Release|x64
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Debug|x64.ActiveCfg = Debug|x64
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Debug|x64.Build.0 = Debug|x64
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Debug|x86.ActiveCfg = Debug|Win32
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Debug|x86.Build.0 = Debug|Win32
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Release|x64.ActiveCfg = Release|x64
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Release|x64.Build.0 = Release|x64
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Release|x86.ActiveCfg = Release|Win32
+		{AF6FBE95-527D-499B-9ABD-3A47E9E84C8A}.Release|x86.Build.0 = Release|Win32
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Debug|x64.ActiveCfg = Debug|x64
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Debug|x64.Build.0 = Debug|x64
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Debug|x86.ActiveCfg = Debug|Win32
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Debug|x86.Build.0 = Debug|Win32
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Release|x64.ActiveCfg = Release|x64
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Release|x64.Build.0 = Release|x64
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Release|x86.ActiveCfg = Release|Win32
+		{05AC9994-2B63-4DE5-ABF3-95CE346F3A64}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {9465DA5B-21F6-46AA-8620-B7544DA98D0F}
+		Qt5Version = msvc2017_64
+	EndGlobalSection
+EndGlobal
