+++ driver_vrmotioncompensation/register_driver.ps1
