+++ driver_vrmotioncompensation/resources/settings/default.vrsettings
