+++ driver_vrmotioncompensation/src/devicemanipulation/DeviceManipulationHandle.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <openvr_driver.h>
-#include <vrinputemulator_types.h>
+#include <vrmotioncompensation_types.h>
 #include "../hooks/common.h"
