+++ driver_vrmotioncompensation/src/devicemanipulation/MotionCompensationManager.cpp
