+++ driver_vrmotioncompensation/src/hooks/ITrackedDeviceServerDriver005Hooks.h
