+++ driver_vrmotioncompensation/src/hooks/IVRServerDriverHost004Hooks.cpp
