+++ driver_vrmotioncompensation/src/hooks/common.cpp
