+++ driver_vrmotioncompensation/src/hooks/common.h
