+++ driver_vrmotioncompensation/src/logging.h
