+++ lib_vrmotioncompensation/src/logging.h
